Use proper type.
Two other places in this file already use uint32_t when passing
it to PCI_READ_LONG.

+++ src/ast_vgatool.c
@@ -3444,7 +3444,7 @@ void ASTGetAST1180DRAMInfo(ScrnInfoPtr pScrn)
 void vASTEnableVGAMMIO(ScrnInfoPtr pScrn)
 {
     ASTRecPtr pAST = ASTPTR(pScrn);
-    ULONG ulData;
+    uint32_t ulData;
     UCHAR jReg;
 
     jReg = inb(pAST->RelocateIO + 0x43);
